Add new landing pages for mobile apps, microservice and serverless
Also improve some of the existing landing pages and update mock up images.
