+++ src/components/shared/MDX.tsx
@@ -7,9 +7,15 @@ import { Button } from 'components/shared/Button';
 import { EmailLink } from 'components/shared/EmailLink';
 import { ExternalLink } from 'components/shared/ExternalLink';
 import { FrameworkSelection } from 'components/shared/FrameworkSelection';
+import { MockRequestsTimeFrame } from 'components/shared/MockRequestsTimeFrame';
+import { PricingCard } from 'components/shared/PricingCard';
 import { SwaggerUI } from 'components/shared/SwaggerUI';
 import { FRONTEND_URL, staticRoutes } from 'utils/router';
 import type { MDXPageProps } from 'types';
+
+const CardContainer: React.FC = ({ children }) => (
+  <div className="rounded-lg border border-base-200 shadow">{children}</div>
+);
 
 const components = {
   Button,
@@ -19,6 +25,16 @@ const components = {
   SwaggerUI,
   Date: ({ date }: { date: string }): string => dayjs(date).format('LL'),
   FrameworkSelection,
+  MockRequestsTimeFrame: (): JSX.Element => (
+    <CardContainer>
+      <MockRequestsTimeFrame />
+    </CardContainer>
+  ),
+  PricingCard: (): JSX.Element => (
+    <CardContainer>
+      <PricingCard />
+    </CardContainer>
+  ),
 };
 
 const scope = {
