+++ src/content/ease-of-use.mdx
@@ -16,7 +16,7 @@ indexable: true
 
 <br />
 
-![Mock-up](/mock-ups/time-frame.gif)
+<MockRequestsTimeFrame />
 
 <br />
