+++ src/content/for-consultants.mdx
@@ -4,7 +4,7 @@ description: Apilytics is ideal for software consultants and freelancers that wa
 indexable: true
 ---
 
-### For consultants 🚀
+### For consultants 👨‍💻
 
 <br />
 
@@ -15,7 +15,7 @@ indexable: true
 
 <br />
 
-![Mock-up](/mock-ups/time-frame.gif)
+<MockRequestsTimeFrame />
 
 <br />
 
@@ -49,4 +49,3 @@ allowing you to easily see how your applications are performing.
 Are you worried about not having the expertise or the right staff to take care of your monitoring?
 Apilytics is built for developers with a mindset that no-one should have hard time understanding their metrics.
 When you choose us, there's no maintainance, no complex configuration and no domain expertise needed.
-Check out our <Link href={routes.demo}>live demo</Link> to learn more!
