+++ src/content/for-microservices.mdx
@@ -1,0 +1,46 @@
+---
+title: For microservices
+description: Apilytics is by far the easiest, yet very scalable APM solution for your microservice architecture with minimal configuration.
+indexable: true
+---
+
+### For microservices 🐳
+
+<br />
+
+<p className="text-xl">
+  Apilytics is by far the easiest, yet very scalable APM (Application Performance Management)
+  solution for your microservice architecture with minimal configuration. The monitoring of your
+  applications does not have to be heavily tied to the infrastructure you have. We achieve this by
+  gaining all of the important metrics from your application directly from the application.
+</p>
+
+<br />
+
+<MockRequestsTimeFrame />
+
+<br />
+
+#### [No host agents](#no-host-agents) {#no-host-agents}
+
+With Apilytics, you can simply embed our open source middleware and there's no need to install any kind of software on the host machines.
+You still get all of the information you need, including performance metrics from your server like CPU and memory usage,
+in addition to the metrics from your applications API calls.
+
+<br />
+
+![Endpoint requests](/mock-ups/endpoint-requests.png)
+
+<br />
+
+#### [Set up once](#set-up-once) {#set-up-once}
+
+Unlike with traditional tools, where you would set up the APM software accross all of your servers and clusters,
+with Apilyitcs the five-minute setup is the only thing you need to do to get started.
+
+<br />
+
+#### [Works everywhere](#works-everywhere) {#works-everywhere}
+
+Apilytics works everywhere, whether your running your applications in public or private clouds or in on-premise setups.
+This way your monitoring infrastructure remains intact even if you make changes to your infrastructure.
