+++ src/content/for-mobile-apps.mdx
@@ -1,0 +1,39 @@
+---
+title: For mobile apps
+description: The application monitoring that Apilytics can do for you covers any kind of clients, including web applications, CLI tools and mobile applications.
+indexable: true
+---
+
+### For mobile apps 📱
+
+<br />
+
+<p className="text-xl">
+  The application monitoring that Apilytics can do for you covers any kind of clients, including web
+  applications, CLI tools and mobile applications. This is why choosing Apilytics is a viable
+  solution to track your mobile app users.
+</p>
+
+<br />
+
+<MockRequestsTimeFrame />
+
+<br />
+
+#### [Track mobile app users separately](#track-mobile-app-users-separately) {#track-mobile-app-users-separately}
+
+Using Apilytics, you can track your mobile app users separately from your web application users.
+Our dashaboard let's you see which type of clients are consuming your APIs.
+By using the user agent filters we provide in our dashaboard, you can easily see stats from your mobile app users.
+
+<br />
+
+![Devices](/mock-ups/user-agent-devices.png)
+
+<br />
+
+#### [Combine metrics with other data](#combine-metrics) {#combine-metrics}
+
+By tracking your API instead of embedding various web analytics tools in your client applications
+that the majority of your users are blocking, you will have combined data accross all of the sources.
+This allows you to easily compare e.g. metrics from your website users with your mobile app users.
