+++ src/content/for-serverless.mdx
@@ -1,0 +1,39 @@
+---
+title: For serverless
+description: Apilytics is the tool you've been waiting for to montor your applications running in serverless environments.
+indexable: true
+---
+
+### For serverless ⚡
+
+<br />
+
+<p className="text-xl">
+  Apilytics is the tool you've been waiting for to montor your applications running in serverless
+  environments. Our way of logging the application metrics directly from an API middleware makes it
+  a perfect solution for any serverless application.
+</p>
+
+<br />
+
+<MockRequestsTimeFrame />
+
+<br />
+
+#### [Monitor individual functions](#monitor-individual-functions) {#monitor-individual-functions}
+
+With Apilytics you have the freedom to monitor individual serverless functions or you can set up
+a broader middleware that monitors all of your functions.
+
+<br />
+
+#### [Works with every cloud provider](#works-with-every-cloud-provider) {#works-with-every-cloud-provider}
+
+You can be sure that Apilytics works in every serverless environment so you will never face a vendor-lock situation.
+
+<br />
+
+#### [Combine with other APIs](#combine-with-other-apis) {#combine-with-other-apis}
+
+If your application is consisting of multiple APIs, some of which might be running in serverless environments,
+you can simply use the same API key to monitor them and you will see metrics from them as separate endpoints in the same dashboard.
