+++ src/content/for-startups.mdx
@@ -4,7 +4,7 @@ description: Apilytics is the optimal monitoring solution for startups. With us,
 indexable: true
 ---
 
-### For startups 🚀
+### For startups 🦄
 
 <br />
 
@@ -17,7 +17,7 @@ indexable: true
 
 <br />
 
-![Mock-up](/mock-ups/time-frame.gif)
+<MockRequestsTimeFrame />
 
 <br />
 
@@ -28,6 +28,10 @@ benefits that Apilytics provides before we charge anything from you.
 After the trial period you can decide if you want to pay for the full service.
 Our pricing model is transparent and easy to understand.
 You will never get unexpected charges for traffic spikes with us.
+
+<br />
+
+<PricingCard />
 
 <br />
