+++ src/content/lightweight.mdx
@@ -15,7 +15,7 @@ indexable: true
 
 <br />
 
-![Mock-up](/mock-ups/time-frame.gif)
+<MockRequestsTimeFrame />
 
 <br />
