+++ src/content/privacy-friendly.mdx
@@ -16,7 +16,7 @@ indexable: true
 
 <br />
 
-![Mock-up](/mock-ups/time-frame.gif)
+<MockRequestsTimeFrame />
 
 <br />
